requirements.txt
Flask
transformers
torch
Pillow
huggingface_hub
